lambda/requirements.txt
autopep8==2.0.2
aws-lambda-powertools==2.18.0
blinker==1.6.2
boto3==1.26.87
botocore==1.29.87
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.3.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
structlog==23.1.0
tomli==2.0.1
typing_extensions==4.6.3
urllib3==1.26.14
Werkzeug==2.3.4
zipp==3.15.0
python-jose==3.3.0